Move sid handling to oauth
