+++ pkg/lib/oauth/oidc/id_token.go
@@ -2,11 +2,8 @@ package oidc
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
 	"net/url"
-	"strings"
-	"unicode/utf8"
 
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jwk"
@@ -51,52 +48,6 @@ type IDTokenIssuer struct {
 // IDTokenValidDuration is the valid period of ID token.
 // It can be short, since id_token_hint should accept expired ID tokens.
 const IDTokenValidDuration = duration.Short
-
-type SessionLike interface {
-	SessionID() string
-	SessionType() session.Type
-}
-
-func EncodeSID(s SessionLike) string {
-	return EncodeSIDByRawValues(s.SessionType(), s.SessionID())
-}
-
-func EncodeSIDByRawValues(sessionType session.Type, sessionID string) string {
-	raw := fmt.Sprintf("%s:%s", sessionType, sessionID)
-	return base64.RawURLEncoding.EncodeToString([]byte(raw))
-}
-
-func DecodeSID(sid string) (typ session.Type, sessionID string, ok bool) {
-	bytes, err := base64.RawURLEncoding.DecodeString(sid)
-	if err != nil {
-		return
-	}
-
-	if !utf8.Valid(bytes) {
-		return
-	}
-	str := string(bytes)
-
-	parts := strings.Split(str, ":")
-	if len(parts) != 2 {
-		return
-	}
-
-	typStr := parts[0]
-	sessionID = parts[1]
-	switch typStr {
-	case string(session.TypeIdentityProvider):
-		typ = session.TypeIdentityProvider
-	case string(session.TypeOfflineGrant):
-		typ = session.TypeOfflineGrant
-	}
-	if typ == "" {
-		return
-	}
-
-	ok = true
-	return
-}
 
 func (ti *IDTokenIssuer) GetPublicKeySet() (jwk.Set, error) {
 	return jwk.PublicSetOf(ti.Secrets.Set)
@@ -330,7 +281,7 @@ func (r *IDTokenHintResolver) ResolveIDTokenHint(ctx context.Context, client *co
 		return
 	}
 
-	typ, sessionID, ok := DecodeSID(sid)
+	typ, sessionID, ok := oauth.DecodeSID(sid)
 	if !ok {
 		return
 	}
